+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\ThreeGoodThings-React\\src\\App.test.js":"1","C:\\ThreeGoodThings-React\\src\\components\\App.js":"2","C:\\ThreeGoodThings-React\\src\\components\\Navbar.js":"3","C:\\ThreeGoodThings-React\\src\\components\\TGT.js":"4","C:\\ThreeGoodThings-React\\src\\components\\TGTInput.js":"5","C:\\ThreeGoodThings-React\\src\\components\\TGTList.js":"6","C:\\ThreeGoodThings-React\\src\\config.js":"7","C:\\ThreeGoodThings-React\\src\\index.js":"8","C:\\ThreeGoodThings-React\\src\\reportWebVitals.js":"9","C:\\ThreeGoodThings-React\\src\\setupTests.js":"10"},{"size":265,"mtime":1605856473621,"results":"11","hashOfConfig":"12"},{"size":3622,"mtime":1606217854455,"results":"13","hashOfConfig":"12"},{"size":2531,"mtime":1606217975685,"results":"14","hashOfConfig":"12"},{"size":1815,"mtime":1606181647050,"results":"15","hashOfConfig":"12"},{"size":5189,"mtime":1606217036861,"results":"16","hashOfConfig":"12"},{"size":2565,"mtime":1606216954869,"results":"17","hashOfConfig":"12"},{"size":432,"mtime":1606210773590,"results":"18","hashOfConfig":"12"},{"size":535,"mtime":1605856648759,"results":"19","hashOfConfig":"12"},{"size":375,"mtime":1605855984234,"results":"20","hashOfConfig":"12"},{"size":246,"mtime":1605855984237,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"oor9pm",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\ThreeGoodThings-React\\src\\App.test.js",[],"C:\\ThreeGoodThings-React\\src\\components\\App.js",[],"C:\\ThreeGoodThings-React\\src\\components\\Navbar.js",[],"C:\\ThreeGoodThings-React\\src\\components\\TGT.js",[],"C:\\ThreeGoodThings-React\\src\\components\\TGTInput.js",[],"C:\\ThreeGoodThings-React\\src\\components\\TGTList.js",[],"C:\\ThreeGoodThings-React\\src\\config.js",[],"C:\\ThreeGoodThings-React\\src\\index.js",[],"C:\\ThreeGoodThings-React\\src\\reportWebVitals.js",[],"C:\\ThreeGoodThings-React\\src\\setupTests.js",[]]
